+++ packages/bluesky/lib/src/entities/actor_profile.g.dart
@@ -24,9 +24,6 @@ _$_ActorProfile _$$_ActorProfileFromJson(Map json) => $checkedCreate(
           postsCount: $checkedConvert('postsCount', (v) => v as int),
           viewer: $checkedConvert('viewer',
               (v) => ActorViewer.fromJson(Map<String, Object?>.from(v as Map))),
-          myState: $checkedConvert('myState',
-              (v) => ActorViewer.fromJson(Map<String, Object?>.from(v as Map))),
-          creator: $checkedConvert('creator', (v) => v as String),
           indexedAt:
               $checkedConvert('indexedAt', (v) => DateTime.parse(v as String)),
         );
@@ -54,8 +51,6 @@ Map<String, dynamic> _$$_ActorProfileToJson(_$_ActorProfile instance) {
   val['followersCount'] = instance.followersCount;
   val['postsCount'] = instance.postsCount;
   val['viewer'] = instance.viewer.toJson();
-  val['myState'] = instance.myState.toJson();
-  val['creator'] = instance.creator;
   val['indexedAt'] = instance.indexedAt.toIso8601String();
   return val;
 }